allow the floating ip to be public ip
For public openstack nova cloud (e.g. trystack.org), floating ip is
available upon request and has to be explicitly attached to the nova
instance, so that user to access (ssh) the nova instance.

Attempt to tweak the existing logics so that the public ip will be
assigned as floating ip. (Existing logics could not do this
automatically)

Tested with trystack.org (public cloud) and walmart openstack (private
cloud)