Fixed issue with backward compatibility
some providers may attach storage to an arbitrary device id, we determine the actual (assigned) device id by issuing attach command and watching a folder on the vm - for instance for azure it's /dev/sd*. Once a new entry appears in that folder - that's our device id.
The problem with the old code is that it's not storing that device id anywhere, and therefore when it runs for the second time it has no way to determine where a particular storage has been attached to.
The new code fixes that by storing device id in a service file (under /opt/oneops/storage_devices).
When we create an assembly using the old code, and then switch to new code and do an update - the update runs but sees no service file, and it tries to simply detach and re-attach the storage. This leads to a problem, because the volume group is already using the storage we're detaching. Fixed by not doing detach