Fixed issue with KCI tests not failing deployments. (#1403)
The culprit is the kitchen-verifier-serverspec gem that provides the serverspec
verifier for KCI tests that we run on inductors (for compute,lb, fqdn, storage,
etc). That verifier executes rspec command using Kernel.system method and does
not check the exit status.
To fix the issue I monkey-patch the Kitchen::Verifier::Serverspec class, to
replace Kernel.system with MixlibL.shellout.