Refactoring of the network recipe in os cookbook.
Removed duplicate functions - in os util library and test util library.
Code cleanup - replacing pure ruby with simple chef resources like file and
template.
Code shared by recipes and kci tests is isolated in functions and moved into a
library.
The main concern was that we had 2 different pieces of code (in the recipe and
in the test library), which would implement the same functionality, just
slightly differently. I tried to re-factor that and to have one implementation
of the functionality (in a library - network_helper.rb), and only have the kci
tests to check the outcome of the function matching the expectation.